Code/requirements.txt
networkx~=3.5
chromadb~=1.0.12
torch_geometric
torch~=2.7.0
node2vec~=0.5.0
neo4j~=5.28.1
openai~=1.84.0
fastapi~=0.115.9
uvicorn
faiss-cpu~=1.11.0
numpy~=1.26.4
sentence-transformers~=4.1.0
pickle-mixin
requests~=2.32.3
asyncpg~=0.30.0
sqlalchemy~=2.0.41
psycopg2~=2.9.10
psycopg2-binary~=2.9.10
python-dotenv~=1.1.0
dotenv
asyncpg
langchain
langchain-core
pydantic
langchain-community>=0.0.20
python-multipart
aiofiles
pillow
google-generativeai
